[integration tests] Move BootstrapConfigTestBuilder in util package
