+++ leshan-integration-tests/src/test/java/org/eclipse/leshan/integration/tests/util/BootstrapConfigTestBuilder.java
@@ -13,7 +13,7 @@
  * Contributors:
  *     Sierra Wireless - initial API and implementation
  *******************************************************************************/
-package org.eclipse.leshan.integration.tests;
+package org.eclipse.leshan.integration.tests.util;
 
 import java.security.PrivateKey;
 import java.security.PublicKey;
@@ -25,8 +25,6 @@ import org.eclipse.leshan.core.SecurityMode;
 import org.eclipse.leshan.core.endpoint.Protocol;
 import org.eclipse.leshan.core.node.LwM2mPath;
 import org.eclipse.leshan.core.request.ContentFormat;
-import org.eclipse.leshan.integration.tests.util.LeshanTestBootstrapServer;
-import org.eclipse.leshan.integration.tests.util.LeshanTestServer;
 import org.eclipse.leshan.server.bootstrap.BootstrapConfig;
 import org.eclipse.leshan.server.bootstrap.BootstrapConfig.ACLConfig;
 import org.eclipse.leshan.server.bootstrap.BootstrapConfig.OscoreObject;
